requirements.txt
llama-index==0.12.44
llama-index-llms-llama-cpp
llama-index-vector-stores-chroma
llama-index-embeddings-openai
chromadb
streamlit
openai
python-dotenv